backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
bleak==0.21.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.8.0
enum-compat==0.0.3
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fsspec==2024.3.1
future==0.18.3
git-filter-repo==2.45.0
git-generator @ file:///Users/addisongoolsbee/Documents/coding/git-generator/dist/git_generator-1.0.0-py3-none-any.whl#sha256=997d9e716fc4279002b45f2a7fce3efddbfd8995aea7b23e41edca2222f11929
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
iso8601==2.1.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.2.3
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
Pygments==2.17.2
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
rich==13.4.2
scikit-learn==1.4.2
scipy==1.13.0
serial==0.0.97
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
typer==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
